pyproject.toml
[project]
name = "uv-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "sqlmodel>=0.0.23",
    "psycopg2-binary==2.9.10",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "sqlalchemy-utils>=0.41.2",
    "pydantic-settings>=2.8.1",
]

[dependency-groups]
dev = []
